Feat/029 relacion products providers (#9)
* Se agrega posibilidad de elegir el proveedor del producto

* Se agregan los proveedores cuando se edita un producto

* Bug fixing en la creacion de provider

* Se agregan los detalles del proveedor a la vista de un producto

* Se agrega el proveedor en el listado de productos y el filtro correspondiente

* refactor products fetchAll

* refactor de las páginas create y update de product

* Se agrega validación para producto

* fix when create producttype

* Se agrega un alert en create product cuando no hay proveedores

* fix. relación obligatoria en productos. problemas al eliminar categoria y proveedor con relacion

* Se agrega valor por defecto en los select de crear producto

* PropType validation en create product

* Se agrega posibilidad de elegir el proveedor del producto

* Se agregan los proveedores cuando se edita un producto

* Bug fixing en la creacion de provider

* Se agregan los detalles del proveedor a la vista de un producto

* Se agrega el proveedor en el listado de productos y el filtro correspondiente

* refactor products fetchAll

* refactor de las páginas create y update de product

* Se agrega validación para producto

* fix when create producttype

* Se agrega un alert en create product cuando no hay proveedores

* fix. relación obligatoria en productos. problemas al eliminar categoria y proveedor con relacion

* Se agrega valor por defecto en los select de crear producto

* PropType validation en create product

* Fix creación de producto

* Fix creación de producto

+++ Stock.Api/Controllers/ProviderController.cs
@@ -10,36 +10,39 @@ using Stock.AppService.Services;
 using Stock.Model.Entities;
 
 namespace Stock.Api.Controllers
-{   
+{
     [Produces("application/json")]
     [Route("api/provider")]
-    [ApiController]    
+    [ApiController]
     public class ProviderController : ControllerBase
     {
         private ProviderService service;
         private readonly IMapper mapper;
-        public ProviderController(ProviderService service, IMapper mapper)
+        private ProductService productService;
+
+        public ProviderController(ProviderService service, IMapper mapper, ProductService productService)
         {
             this.service = service;
             this.mapper = mapper;
+            this.productService = productService;
         }
 
         [HttpPost]
         public ActionResult Post([FromBody] ProviderDTO value)
         {
             TryValidateModel(value);
-            
+
             try
             {
                 var provider = this.mapper.Map<Provider>(value);
                 this.service.Create(provider);
                 value.Id = provider.Id;
-                return Ok(new {Success = true, Message = "", data = value});
+                return Ok(new { Success = true, Message = "", data = value });
             }
-            catch            
+            catch
             {
-                return Ok(new {Success = false, Message = "The name is already in use"});
-            }        
+                return Ok(new { Success = false, Message = "The name is already in use" });
+            }
         }
 
         [HttpGet]
@@ -50,7 +53,7 @@ namespace Stock.Api.Controllers
                 var result = this.service.GetAll();
                 return this.mapper.Map<IEnumerable<ProviderDTO>>(result).ToList();
             }
-            catch(Exception)
+            catch (Exception)
             {
                 return StatusCode(500);
             }
@@ -64,7 +67,7 @@ namespace Stock.Api.Controllers
                 var result = this.service.Get(id);
                 return this.mapper.Map<ProviderDTO>(result);
             }
-            catch(Exception)
+            catch (Exception)
             {
                 return StatusCode(500);
             }
@@ -84,12 +87,23 @@ namespace Stock.Api.Controllers
         /// </summary>
         /// <param name="id">Identificador de la instancia a borrar</param>
         [HttpDelete("{id}")]
-        public void Delete(string id)
+        public ActionResult Delete(string id)
         {
             var provider = this.service.Get(id);
+
+            Expression<Func<Product, bool>> filter = x => x.ProviderId.Equals(id);
+
+            var products = this.productService.Search(filter);
+
+            if (products.Count() > 0)
+            {
+               return Ok(new { Success = false, Message = "No se pueden eliminar proveedores que tiene productos asociados" });
+            }
+
             this.service.Delete(provider);
+            return Ok();
         }
-        
+
         [HttpPost("/search")]
         public ActionResult Search([FromBody] ProviderSearchDTO model)
         {
@@ -102,7 +116,7 @@ namespace Stock.Api.Controllers
                     model.Condition.Equals(ActionDto.AND));
             }
 
-            if(!string.IsNullOrWhiteSpace(model.Email))
+            if (!string.IsNullOrWhiteSpace(model.Email))
             {
                 filter = filter.AndOrCustom(
                     x => x.Email.ToUpper().Contains(model.Email.ToUpper()),
